* ConsoleRunner: fix for WIN64 - cast command return code to INT
Ensure on WIN64 systems that the status code displayed in the frame 'Build log'
is identical to the return code of the executable run by cb_console_runner.exe.

+++ src/tools/ConsoleRunner/main.cpp
@@ -68,7 +68,7 @@ int execute_command(char *cmdline)
     CloseHandle( pi.hProcess );
     CloseHandle( pi.hThread );
 
-    return ret;
+    return (INT)ret;
 #else
     int ret = system(cmdline);
     if(WIFEXITED(ret))
